Patch for decimal heights (due to Tiled editor) and ability to load external tilesets.

+++ src/NTiled/Domain/TiledObject.cs
@@ -59,13 +59,13 @@ namespace NTiled
         /// Gets or sets the width of the object in pixels.
         /// </summary>
         /// <value>The width of the object in pixels.</value>
-        public int Width { get; set; }
+        public decimal Width { get; set; }
 
         /// <summary>
         /// Gets or sets the height of the object in pixels.
         /// </summary>
         /// <value>The height of the object in pixels.</value>
-        public int Height { get; set; }
+        public decimal Height { get; set; }
 
         /// <summary>
         /// Gets the object properties.
